+++ start/src/App.vue
@@ -1,5 +1,34 @@
-<script lang="ts" setup></script>
+<script setup lang="ts">
+  import { ref } from 'vue'
 
-<template></template>
+  const visible1 = ref(false)
+</script>
 
-<style lang="scss" scoped></style>
+<template>
+  <f-button type="primary" @click="visible1 = true">打开</f-button>
+
+  <f-dialog v-model:visible="visible1" title="标题文字">
+    这是一个 Dialog
+    <h1>11</h1>
+    <h1>11</h1>
+    <h1>11</h1>
+    <h1>11</h1>
+    <h1>11</h1>
+    <h1>11</h1>
+    <h1>11</h1>
+    <h1>11</h1>
+    <h1>11</h1>
+    <h1>11</h1>
+    <h1>11</h1>
+    <h1>11</h1>
+    <h1>11</h1>
+    <h1>11</h1>
+    <h1>11</h1>
+    <h1>11</h1>
+
+    <template #footer>
+      <f-button type="default">默认按钮</f-button>
+      <f-button type="primary">主要按钮</f-button>
+    </template>
+  </f-dialog>
+</template>